fix(dashboard-query): separate dashboard query context
